remove controller package and move code up a level
